add default profile picture, fix filter and navbar
