+++ home.html
@@ -24,7 +24,7 @@
 
     <img
       id="userAvatar"
-      src=""
+      src="./images/profilePicture.png"
       alt="User Avatar"
       style="width: 40px; height: 40px; border-radius: 50%"
     />
