+++ src/components/site-navbar.js
@@ -1,27 +1,36 @@
 // Import specific functions from the Firebase Auth SDK
 import { onAuthStateChanged } from "firebase/auth";
-
-import { auth } from "/src/firebaseConfig.js";
+import { doc, getDoc } from "firebase/firestore";
+import { auth, db } from "/src/firebaseConfig.js";
 import { logoutUser } from "/src/authentication.js";
 
 class SiteNavbar extends HTMLElement {
   constructor() {
     super();
     this.renderNavbar();
-    this.renderAuthControls();
+    this.setupAuthListener();
   }
 
   renderNavbar() {
     this.innerHTML = `
   <nav class="navbar bg-body-tertiary bg-info">
     <div class="container-fluid">
-      <a class="navbar-brand" href="/">
-        <img src="/images/logo.png" height="36">
-         
-      </a>
-      <button class="navbar-toggler" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasNavbar" aria-controls="offcanvasNavbar" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
+      <div class="d-flex align-items-center">
+        <button class="navbar-toggler me-2" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasNavbar" aria-controls="offcanvasNavbar" aria-label="Toggle navigation">
+          <span class="navbar-toggler-icon"></span>
+        </button>
+        <a class="navbar-brand" href="/home.html">
+          <img src="/images/logo.png" height="36">
+        </a>
+      </div>
+
+      <img
+        id="userAvatar"
+        src="/images/profilePicture.png"
+        alt="User Avatar"
+        style="width: 40px; height: 40px; border-radius: 50%; object-fit: cover;"
+      />
+
       <div class="offcanvas offcanvas-end w-75" tabindex="-1" id="offcanvasNavbar" aria-labelledby="offcanvasNavbarLabel">
         <div class="offcanvas-header">
           <h5 class="offcanvas-title" id="offcanvasNavbarLabel">FlipIt</h5>
@@ -50,24 +59,73 @@ class SiteNavbar extends HTMLElement {
   `;
   }
 
-  renderAuthControls() {
-    const authControls = this.querySelector("#authControls");
-
-    // Initialize with invisible placeholder to maintain layout space
-    authControls.innerHTML = `<div class="btn btn-outline-light" style="visibility: hidden; min-width: 80px;">Log out</div>`;
-
-    onAuthStateChanged(auth, (user) => {
-      let updatedAuthControl;
+  setupAuthListener() {
+    onAuthStateChanged(auth, async (user) => {
       if (user) {
-        updatedAuthControl = `<button class="btn btn-outline-dark" id="signOutBtn" type="button" style="min-width: 80px;">Log out</button>`;
-        authControls.innerHTML = updatedAuthControl;
-        const signOutBtn = authControls.querySelector("#signOutBtn");
-        signOutBtn?.addEventListener("click", logoutUser);
+        console.log("Get the ID:", user.uid);
+
+        try {
+          const userRef = doc(db, "users", user.uid);
+          const userSnap = await getDoc(userRef);
+
+          if (userSnap.exists()) {
+            const userData = userSnap.data();
+            console.log("user data:", userData);
+
+            if (userData.avatarUrl) {
+              console.log("Find the URL:", userData.avatarUrl);
+              const avatarElement = this.querySelector("#userAvatar");
+
+              if (avatarElement) {
+                avatarElement.src = userData.avatarUrl;
+                console.log("The picture already set");
+              } else {
+                console.log("Can't find the picture element in navbar");
+              }
+            } else {
+              console.log("The user didn't set the picture");
+              const avatarElement = this.querySelector("#userAvatar");
+              if (avatarElement) {
+                avatarElement.src = "/images/profilePicture.png";
+              }
+            }
+          } else {
+            console.log("The user document does not exist");
+          }
+        } catch (error) {
+          console.error("Error in obtaining user data:", error);
+        }
+
+        this.updateAuthControls(true);
       } else {
-        updatedAuthControl = `<a class="btn btn-outline-dark" id="loginBtn" href="/login.html" style="min-width: 80px;">Log in</a>`;
-        authControls.innerHTML = updatedAuthControl;
+        console.log("The user is not logged in");
+        this.setDefaultAvatar();
+        this.updateAuthControls(false);
       }
     });
+  }
+
+  setDefaultAvatar() {
+    const avatarElement = this.querySelector("#userAvatar");
+    if (avatarElement) {
+      avatarElement.src = "/images/profilePicture.png";
+    }
+  }
+
+  updateAuthControls(isLoggedIn) {
+    const authControls = this.querySelector("#authControls");
+    if (!authControls) return;
+
+    let updatedAuthControl;
+    if (isLoggedIn) {
+      updatedAuthControl = `<button class="btn btn-outline-dark" id="signOutBtn" type="button" style="min-width: 80px;">Log out</button>`;
+      authControls.innerHTML = updatedAuthControl;
+      const signOutBtn = authControls.querySelector("#signOutBtn");
+      signOutBtn?.addEventListener("click", logoutUser);
+    } else {
+      updatedAuthControl = `<a class="btn btn-outline-dark" id="loginBtn" href="/login.html" style="min-width: 80px;">Log in</a>`;
+      authControls.innerHTML = updatedAuthControl;
+    }
   }
 }
