+++ src/styles/home.css
@@ -327,12 +327,12 @@ site-navbar {
 
 #userAvatar {
   position: fixed;
-  top: 10px;
-  right: 100px;
-  width: 40px;
-  height: 40px;
+  top: 15px;
+  right: 20px;
+  width: 60px;
+  height: 60px;
   border-radius: 50%;
   z-index: 1001;
   border: 2px solid white;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
+  /* box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2); */
 }
